requirements.txt
PyYAML==3.12
async-timeout==1.2.1
asyncio==3.4.3
discord.py==0.16.8
pytest==3.1.1